micapp: Audio recording effects
1)
Added a check and control of input effects i.e.
* Acoustic echo canceler (AEC)
* Automatic gain control (AGC)
* Noise supression (NS)

When a input is selected and the active button is on the stqatus of
the effects will be updated and if they are not available they will
have a red tinted bckground.

If available but not enabled it will be a gray background saying:
"No agc" etc.

2)
If recording is enabled the recording path wil be show in the info
panel together with information about the used input.

3)
Spl metering is showing
current (max / min)
in dB, for both peak and rms.
Measuring period is one second.

Problem:
* It seesm like the effects can be slow if creted/destroyed
rapidly and not be turned on sometimes.